backend/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
google-generativeai==0.8.3
qdrant-client==1.12.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.3
openai-agents>=0.6.1
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator

